feat: add optimizations for CUDA
This commit adds optimizations for CUDA based on the work from Supranational.
These optimizations are only enabled on CUDA, as it increases the compile time
on OpenCL too much to be practical. On CUDA the kernels will be pre-compiled
at compile time to work around that problem (also the CUDA compile time is
way shorter than on OpenCL).